+++ packages/core-js/internals/get-regexp-flags-implementation.js
