+++ packages/core-js/modules/es.regexp.constructor.js
@@ -10,7 +10,7 @@ var getOwnPropertyNames = require('../internals/object-get-own-property-names').
 var isPrototypeOf = require('../internals/object-is-prototype-of');
 var isRegExp = require('../internals/is-regexp');
 var toString = require('../internals/to-string');
-var getRegExpFlagsImplementation = require('../internals/regexp-get-flags');
+var getRegExpFlagsImplementation = require('../internals/get-regexp-flags-implementation');
 var stickyHelpers = require('../internals/regexp-sticky-helpers');
 var proxyAccessor = require('../internals/proxy-accessor');
 var defineBuiltIn = require('../internals/define-built-in');
