+++ packages/core-js/modules/es.regexp.flags.js
@@ -1,10 +1,10 @@
 'use strict';
-var checkCorrectnessOfRegExpFlags = require('../internals/check-correctness-of-regexp-flags');
-var detectionOfCorrectnessRegExpFlags = require('../internals/detection-of-correctness-regexp-flags');
+var DESCRIPTORS = require('../internals/descriptors');
+var detectCorrectnessOfRegExpFlags = require('../internals/detect-correctness-of-regexp-flags');
 var defineBuiltInAccessor = require('../internals/define-built-in-accessor');
 var regExpFlags = require('../internals/regexp-flags');
 
-var FORCED = !checkCorrectnessOfRegExpFlags();
+var FORCED = DESCRIPTORS && !detectCorrectnessOfRegExpFlags.correct;
 
 // `RegExp.prototype.flags` getter
 // https://tc39.es/ecma262/#sec-get-regexp.prototype.flags
@@ -13,5 +13,5 @@ if (FORCED) {
     configurable: true,
     get: regExpFlags
   });
-  detectionOfCorrectnessRegExpFlags.correct = true;
+  detectCorrectnessOfRegExpFlags.correct = true;
 }
